+++ src/domain/use_cases/pedido/pedido.use_case.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { IPedidoRepository } from 'src/domain/ports/pedido/pedido.repository.port';
 import { IPedidoFactory } from 'src/domain/ports/pedido/pedido.factory.port';
-import { pedidoRepositoryMock } from 'src/mocks/pedido/pedido.repository.mock';
+import { pedidoRepositoryMock } from 'src/mocks/pedido.repository.mock';
 import { pedidoFactoryMock } from 'src/mocks/pedido/pedido.factory.mock';
 import { PedidoUseCase } from './pedido.use_case';
