+++ src/mocks/categoria.mock.ts
@@ -1,0 +1,19 @@
+import { CategoriaModel } from 'src/adapters/outbound/models/categoria.model';
+import { CategoriaEntity } from 'src/domain/entities/categoria.entity';
+
+const categoriaEntity = new CategoriaEntity(
+  'Lanche',
+  'Lanche x tudo',
+  '0a14aa4e-75e7-405f-8301-81f60646c93d',
+);
+
+const categoriaModel = new CategoriaModel();
+categoriaModel.id = '0a14aa4e-75e7-405f-8301-81f60646c93d';
+categoriaModel.nome = 'Lanche';
+categoriaModel.descricao = 'Lanche x tudo';
+categoriaModel.produtos = null;
+categoriaModel.criadoEm = new Date().toISOString();
+categoriaModel.atualizadoEm = new Date().toISOString();
+categoriaModel.excluidoEm = new Date().toISOString();
+
+export { categoriaEntity, categoriaModel };
