--- src/mocks/pedido/pedido.service.mock.ts
@@ -1,8 +1,0 @@
-import { PedidoService } from 'src/domain/services/pedido.service';
-
-const pedidoServiceMock: jest.Mocked<Pick<PedidoService, 'gerarNumeroPedido'>> =
-  {
-    gerarNumeroPedido: jest.fn(),
-  };
-
-export { pedidoServiceMock };
